+++ src/OpenTelemetry.Exporter.Console/CHANGELOG.md
@@ -6,6 +6,12 @@ Notes](../../RELEASENOTES.md).
 
 ## Unreleased
 
+* Added support for `ActivitySource.TelemetrySchemaUrl` property.
+  ([#6713](https://github.com/open-telemetry/opentelemetry-dotnet/pull/6713))
+
+* Added support for `Meter.TelemetrySchemaUrl` property.
+  ([#6714](https://github.com/open-telemetry/opentelemetry-dotnet/pull/6714))
+
 ## 1.14.0
 
 Released 2025-Nov-12