+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -12,6 +12,11 @@ Notes](../../RELEASENOTES.md).
   the custom identifier is prepended with a space separator to the default
   User-Agent string (e.g., `MyApp/1.0 OTel-OTLP-Exporter-Dotnet/1.14.0`).
   ([#6686](https://github.com/open-telemetry/opentelemetry-dotnet/pull/6686))
+* Added support for `ActivitySource.TelemetrySchemaUrl` property.
+  ([#6730](https://github.com/open-telemetry/opentelemetry-dotnet/pull/6730))
+
+* Added support for `Meter.TelemetrySchemaUrl` property.
+  ([#6731](https://github.com/open-telemetry/opentelemetry-dotnet/pull/6731))
 
 ## 1.14.0