+++ src/OpenTelemetry/CHANGELOG.md
@@ -6,6 +6,12 @@ Notes](../../RELEASENOTES.md).
 
 ## Unreleased
 
+* Added `LowMemory` temporality as an option in the OTLP metrics exporter.
+  ([#6648](https://github.com/open-telemetry/opentelemetry-dotnet/pull/6648))
+
+* Added support for `Meter.TelemetrySchemaUrl` property.
+  ([#6714](https://github.com/open-telemetry/opentelemetry-dotnet/pull/6714))
+
 ## 1.14.0
 
 Released 2025-Nov-12